+++ src/Compilers/ContainerBuilder.php
@@ -15,7 +15,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace BiuradPHP\DependencyInjection\Concerns;
+namespace BiuradPHP\DependencyInjection\Compilers;
 
 use BiuradPHP\DependencyInjection\Container;
 use BiuradPHP\DependencyInjection\Definitions\InterfaceDefinition;
@@ -36,8 +36,7 @@ use function BiuradPHP\Support\array_get;
 /**
  * Container builder.
  *
- * @author David Grudl <https://davidgrudl.com>
- * @license BSD-3-Clause
+ * @author Divine Niiquaye Ibok <divineibok@gmail.com>
  */
 class ContainerBuilder extends NetteContainerBuilder
 {