+++ src/Singleton.php
@@ -18,9 +18,14 @@ declare(strict_types=1);
 namespace BiuradPHP\DependencyInjection;
 
 /**
- * The singleton base class restricts the instantiation of derived classes to one object only.
+ * The singleton base class restricts the instantiation of
+ * derived classes to one object only.
+ *
+ * @final
+ *
+ * @author Divine Niiquaye Ibok <divineibok@gmail.com>
  */
-final class Singleton
+class Singleton
 {
     /**
      * @var Singleton