+++ src/Traits/ContainerAwareTrait.php
@@ -15,7 +15,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace BiuradPHP\DependencyInjection\Concerns;
+namespace BiuradPHP\DependencyInjection\Traits;
 
 use Psr\Container\ContainerInterface;