+++ CHANGELOG.md
@@ -5,8 +5,17 @@ All notable changes to `biurad/nette-di-bridge` will be documented in this file.
 ## 0.1.7 - 2020-06-17
 
 - Improved code complexity and performance using cs fixtures
+- Added service definitions access from xml files
+- Added `BiuradPHP\DependencyInjection\Interfaces\ResolverInterface`
+- Added `BiuradPHP\DependencyInjection\Interfaces\ScopeInterface`
+- Added annotations prototype feature to container
+- Added phpunit tests
+- Removed `bindMethod` from DependencyInjection container
+- Replaced `InvalidStateException` with PSR-11 `ContainerResolutionException`
+- Moved php class files from `Concerns` to `Compilers` folder
+- Marked `Singleton` class as final
 - Updated php files header doc
-- Added phpunit and nette tests
+- Updated README.md file
 
 ## 0.1.0 - 2020-04-26